requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.5
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
Markdown==3.5.1
pycparser==2.21
pytz==2023.3.post1
sqlparse==0.4.4
